fix: merge schemas
