+++ src/actions/getSquares.ts
@@ -1,189 +1,17 @@
-import { auth } from "@/actions/auth";
-import { Square } from "@/models/Square";
+"use server";
 
-export async function getSquares(): Promise<Square[]> {
+import "server-only";
+
+import { auth } from "@/actions/auth";
+import { Board } from "@/models/Board";
+import { getTeamBoard } from "@/services/teamBoard";
+
+export async function getBoard(): Promise<Board> {
   const { teamId } = await auth();
 
   if (!teamId) {
     throw new Error("Team ID is required");
   }
 
-  return [
-    {
-      complete: false,
-      points: 1,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 0,
-        y: 0,
-      },
-    },
-    {
-      complete: false,
-      points: 2,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 0,
-        y: 1,
-      },
-    },
-    {
-      complete: false,
-      points: 3,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 0,
-        y: 2,
-      },
-    },
-    {
-      complete: false,
-      points: 4,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 0,
-        y: 3,
-      },
-    },
-    {
-      complete: false,
-      points: 2,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 1,
-        y: 0,
-      },
-    },
-    {
-      complete: false,
-      points: 3,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 1,
-        y: 1,
-      },
-    },
-    {
-      complete: false,
-      points: 4,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 1,
-        y: 2,
-      },
-    },
-    {
-      complete: false,
-      points: 1,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 1,
-        y: 3,
-      },
-    },
-    {
-      complete: false,
-      points: 3,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 2,
-        y: 0,
-      },
-    },
-    {
-      complete: false,
-      points: 4,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 2,
-        y: 1,
-      },
-    },
-    {
-      complete: false,
-      points: 1,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 2,
-        y: 2,
-      },
-    },
-    {
-      complete: false,
-      points: 2,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 2,
-        y: 3,
-      },
-    },
-    {
-      complete: false,
-      points: 4,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 3,
-        y: 0,
-      },
-    },
-    {
-      complete: false,
-      points: 1,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 3,
-        y: 1,
-      },
-    },
-    {
-      complete: false,
-      points: 2,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 3,
-        y: 2,
-      },
-    },
-    {
-      complete: false,
-      points: 3,
-      activity: {
-        id: "test",
-        name: "test",
-        description: "test",
-        x: 3,
-        y: 3,
-      },
-    },
-  ];
+  return await getTeamBoard(teamId);
 }
