+++ src/actions/getTeamPoints.ts
@@ -1,19 +1,17 @@
 "use server";
+
 import { getTeamBoard } from "@/services/teamBoard";
 import "server-only";
 
 export async function getTeamPoints(teamId: string) {
-    const board = await getTeamBoard(teamId)
+  const board = await getTeamBoard(teamId);
 
-    let points = 0;
-    board.forEach(row => {
-        row.forEach(square => {
-            if (square.completed) {
-                points += square.points;
-            }
-        });
-    });
+  let points = 0;
+  board.forEach((square) => {
+    if (square.complete) {
+      points += square.points;
+    }
+  });
 
-    return points;
-
+  return points;
 }
