+++ src/models/Board.ts
@@ -1,18 +1,22 @@
 import { z } from "zod";
 
-import { ActivitySchema } from "./Activity";
-
 /** A single square in the board */
-export const SquareSchema = z.object({
-  completed: z.boolean({ message: "Square has invalid completed" }),
-  activity: ActivitySchema,
-  points: z.number({ message: "Square has invalid points" }),
+export const ActivitySchema = z.object({
+  id: z.string(),
+  name: z.string(),
+  description: z.string(),
+  x: z.number(),
+  y: z.number(),
 });
 
-/** 4x4 board */
-export const BoardSchema = z
-  .array(z.array(SquareSchema).length(4, { message: "Invalid board" }))
-  .length(4, { message: "Invalid board" });
+export const SquareSchema = z.object({
+  complete: z.boolean(),
+  points: z.number(),
+  activity: ActivitySchema,
+});
 
+export const BoardSchema = z.array(SquareSchema);
+
+export type Activity = z.infer<typeof ActivitySchema>;
 export type Square = z.infer<typeof SquareSchema>;
 export type Board = z.infer<typeof BoardSchema>;
