+++ src/queries/useGetBoard.ts
@@ -1,14 +1,12 @@
-import { z } from "zod";
-
-import { getSquares } from "@/actions/getSquares";
-import { SquareSchema } from "@/models/Square";
+import { getBoard } from "@/actions/getSquares";
+import { BoardSchema } from "@/models/Board";
 import { useSWRWithZod } from "@/utils/swr";
 
 export default function useGetBoard() {
   return useSWRWithZod({
     cacheKey: `useGetSquares`, //TODO: somehow get the teamid in here
-    fetcher: getSquares,
-    zodSchema: z.array(SquareSchema),
+    fetcher: getBoard,
+    zodSchema: BoardSchema,
     thisFile: "queries/useGetSquares.ts",
   });
 }
